board_enable.c: Add ME unlock function for TUXEDO laptops
Add board enable function that unlocks the ME region with EC GPIO
assistance which asserts the FD security override strap on the next
power off and power on cycle. The pin is asserted only for one boot
duration until next reset. There is a timer defined for the S5 state
residence time and programmed to EC to automatically power on the
platform back when the operation is successful.

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Change-Id: I6bf1c40900aa674e3ea4f6add12dae8b73759fbb